--- a/setup.cfg
+++ b/setup.cfg
@@ -1,7 +1,9 @@
 [flake8]
 max-line-length = 100
+ignore = E704,W503,PYD002,TC003,TC006
 exclude =
 	venv*
 	.eggs
 	docs
 	build
+type-checking-pydantic-enabled = True