### What I did

before:

```
In [3]: %time import ape_polygon_zkevm
exitCPU times: user 1.74 s, sys: 159 ms, total: 1.9 s
Wall time: 1.9 s
```

after:

```
CPU times: user 7.84 ms, sys: 3.27 ms, total: 11.1 ms
Wall time: 10.4 ms
```

### How I did it

### How to verify it

### Checklist

- [ ] Passes all linting checks (pre-commit and CI jobs)
- [ ] New test cases have been added and are passing
- [ ] Documentation has been updated
- [ ] PR title follows [Conventional Commit](https://www.conventionalcommits.org/en/v1.0.0/) standard (will be automatically included in the changelog)